+++ test/clj/lambdacd/steps/result_test.clj
@@ -44,12 +44,4 @@
     (is (= nil (join-output-resolver :other-key "foo" :bar)))))
 
 
-(deftest flatten-step-result-outputs-test
-  (testing "that it works"
-    (is (= {[1]   {:status :success}
-            [2]   {:status  :success
-                   :outputs {[1 2] {:status :success :step [1 2]}}}
-            [1 2] {:status :success :step [1 2]}}
-           (flatten-step-result-outputs {[1] {:status :success}
-                                         [2] {:status  :success
-                                              :outputs {[1 2] {:status :success :step [1 2]}}}})))))
+